--- src/main/java/edu/ithaca/dturnbull/bank/BankTeller.java
@@ -1,29 +1,0 @@
-package edu.ithaca.dturnbull.bank;
-
-
-public class BankTeller extends AbstractTeller {
-
-    
-    public BankTeller(){
-
-    }
-
-    @Override
-    public boolean login(String username, String password){
-        if (this.username.equals(username)){
-            if (this.password.equals(password)){
-                return true;
-            }
-            else{
-                return false;
-            }
-        }
-        return false;
-    }
-
-    @Override
-    public void createAccount(){
-        
-    }
-    
-}
